feat: improve reservation search feedback in UI
This commit improves the user interface for reservation searches. It adds a more descriptive message when a reservation search yields no results, and displays a different message to prompt the user to perform a search when the page is first loaded. Also, a small style adjustment was made to increase the bottom margin of the return button. The changes also involve adjusting the reservation controller to pass an additional variable to the template to indicate whether the search form has been submitted.